[LXC] Scope PR to most-specific-path resolver; drop file/dir schema work
Address review feedback on #630:

- Remove the denied-path file/dir `type` discriminator (wire DeniedPath
  union, MaskKind, denied_path_kinds, regenerated schema + TS wire types,
  ts_emit anyOf/allOf handling, and the LXC mask-kind logic). Per reviewer
  guidance the file-vs-dir schema distinction is deferred (to be designed
  alongside leaf/tree), and the wire change duplicates bwrap #640.
- Keep the most-specific-path-wins resolver but extend `filesystem_resolve`
  in place instead of renaming it to `path_specificity` with a compat shim,
  avoiding churn / conflict surface with the bwrap branches. Delete
  path_specificity.rs and the shim.
- Fix the `dedup_by` dead code: the retained element is the earlier `b`, so
  mutating `a` was a no-op. Rely on the most-restrictive-first sort (with a
  comment) and add a test asserting the kept path string on an exact-key
  conflict across lists that differ only by a trailing slash.
- Revert LXC denied-path masking to its original behavior while still
  emitting mounts in resolver (shallow->deep) order.
