+++ src/backends/lxc/common/src/filesystem_mounts.rs
@@ -9,12 +9,9 @@
 //! - `readonlyPaths` → `bind,ro` mount
 //! - `deniedPaths` → masked (inaccessible inside container)
 
-use std::collections::HashMap;
-use std::path::Path;
-
+use wxc_common::filesystem_resolve::{resolve_mount_order, FsIntent};
 use wxc_common::logger::Logger;
-use wxc_common::models::{ContainerPolicy, MaskKind};
-use wxc_common::path_specificity::{resolve_mount_order, FsIntent};
+use wxc_common::models::ContainerPolicy;
 
 use crate::lxc_bindings::LxcContainer;
 
@@ -33,92 +30,6 @@ fn validate_path(path: &str) -> Result<(), String> {
     }
 
     Ok(())
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-enum ObservedMaskPathKind {
-    File,
-    Dir,
-    Symlink,
-}
-
-/// Inspect a denied path on the **host** filesystem to infer its mask kind.
-///
-/// Uses `symlink_metadata` so a symlink is classified as itself rather than
-/// followed (avoids the TOCTOU-prone `is_file` heuristic). A missing path
-/// yields `Ok(None)` so the caller can fall back to an explicit `type`.
-///
-/// This must observe the host path — not the container rootfs path, which does
-/// not exist before mounts are applied — otherwise every host denied path would
-/// resolve to `NotFound` and spuriously demand an explicit `type`.
-fn observed_mask_path_kind(host_path: &Path) -> Result<Option<ObservedMaskPathKind>, String> {
-    match std::fs::symlink_metadata(host_path) {
-        Ok(metadata) => {
-            let file_type = metadata.file_type();
-            if file_type.is_dir() {
-                Ok(Some(ObservedMaskPathKind::Dir))
-            } else if file_type.is_symlink() {
-                Ok(Some(ObservedMaskPathKind::Symlink))
-            } else {
-                Ok(Some(ObservedMaskPathKind::File))
-            }
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(None),
-        Err(err) => Err(format!(
-            "Unable to inspect denied path '{}': {}. Set deniedPaths entry to an object with explicit type \"file\" or \"dir\".",
-            host_path.display(),
-            err
-        )),
-    }
-}
-
-/// Trim trailing path separators from a denied-path key so an explicit `type`
-/// is not dropped when the resolved mount path and the configured
-/// `denied_path_kinds` key differ only by a trailing slash. A root-only path is
-/// preserved rather than normalized to an empty string.
-fn normalize_denied_key(path: &str) -> &str {
-    let trimmed = path.trim_end_matches(['/', '\\']);
-    if trimmed.is_empty() {
-        path
-    } else {
-        trimmed
-    }
-}
-
-/// Look up the explicit mask kind for a denied host path, tolerating a
-/// trailing-separator mismatch between the resolved mount path and the
-/// configured key. The exact match is tried first as a fast path.
-fn lookup_mask_kind(
-    denied_path_kinds: &HashMap<String, MaskKind>,
-    host_path: &str,
-) -> Option<MaskKind> {
-    if let Some(kind) = denied_path_kinds.get(host_path) {
-        return Some(*kind);
-    }
-    let normalized = normalize_denied_key(host_path);
-    denied_path_kinds
-        .iter()
-        .find(|(key, _)| normalize_denied_key(key) == normalized)
-        .map(|(_, kind)| *kind)
-}
-
-fn resolve_mask_kind(
-    denied_path: &str,
-    explicit: Option<MaskKind>,
-    observed: Option<ObservedMaskPathKind>,
-) -> Result<MaskKind, String> {
-    if let Some(kind) = explicit {
-        return Ok(kind);
-    }
-
-    match observed {
-        Some(ObservedMaskPathKind::Dir) => Ok(MaskKind::Dir),
-        Some(ObservedMaskPathKind::File | ObservedMaskPathKind::Symlink) => Ok(MaskKind::File),
-        None => Err(format!(
-            "Denied path '{}' does not exist and no mask type was specified. Use deniedPaths object form {{\"path\":\"{}\",\"type\":\"file\"}} or {{\"path\":\"{}\",\"type\":\"dir\"}}.",
-            denied_path, denied_path, denied_path
-        )),
-    }
 }
 
 /// Configure filesystem bind mounts on the container based on the policy.
@@ -156,25 +67,20 @@ pub fn configure_filesystem_mounts(
                 container.set_config_item("lxc.mount.entry", &mount_entry)?;
             }
             FsIntent::Denied => {
-                let explicit = lookup_mask_kind(&policy.denied_path_kinds, host_path);
-                let kind = resolve_mask_kind(
-                    host_path,
-                    explicit,
-                    observed_mask_path_kind(Path::new(host_path))?,
-                )?;
+                // Determine if the path is a file or directory inside the
+                // container rootfs so we use the correct LXC `create=` type for
+                // the mount entry.
+                let rootfs_base = format!("{}/{}/rootfs", container.lxc_path(), container.name());
+                let full_path = format!("{}/{}", rootfs_base, container_path);
 
-                let mount_entry = match kind {
-                    MaskKind::File => {
-                        format!("/dev/null {} none bind,ro,create=file 0 0", container_path)
-                    }
-                    MaskKind::Dir => {
-                        format!("tmpfs {} tmpfs ro,size=0,create=dir 0 0", container_path)
-                    }
+                // Use /dev/null bind mount for files, tmpfs for directories.
+                let is_file = std::path::Path::new(&full_path).is_file();
+                let mount_entry = if is_file {
+                    format!("/dev/null {} none bind,ro,create=file 0 0", container_path)
+                } else {
+                    format!("tmpfs {} tmpfs ro,size=0,create=dir 0 0", container_path)
                 };
-                let create_type = match kind {
-                    MaskKind::File => "file",
-                    MaskKind::Dir => "dir",
-                };
+                let create_type = if is_file { "file" } else { "dir" };
                 logger.log_line(&format!(
                     "Masking denied path: /{} ({})",
                     container_path, create_type
@@ -233,124 +139,5 @@ mod tests {
     #[test]
     fn test_validate_path_accepts_normal() {
         assert!(validate_path("/mnt/shared").is_ok());
-    }
-
-    #[test]
-    fn explicit_mask_kind_wins_over_observed_kind() {
-        assert_eq!(
-            resolve_mask_kind(
-                "/etc/shadow",
-                Some(MaskKind::File),
-                Some(ObservedMaskPathKind::Dir)
-            )
-            .unwrap(),
-            MaskKind::File
-        );
-        assert_eq!(
-            resolve_mask_kind(
-                "/var/lib/app",
-                Some(MaskKind::Dir),
-                Some(ObservedMaskPathKind::File)
-            )
-            .unwrap(),
-            MaskKind::Dir
-        );
-    }
-
-    #[test]
-    fn observed_symlink_and_regular_file_use_file_mask() {
-        assert_eq!(
-            resolve_mask_kind(
-                "/etc/alternatives/editor",
-                None,
-                Some(ObservedMaskPathKind::Symlink)
-            )
-            .unwrap(),
-            MaskKind::File
-        );
-        assert_eq!(
-            resolve_mask_kind("/etc/shadow", None, Some(ObservedMaskPathKind::File)).unwrap(),
-            MaskKind::File
-        );
-    }
-
-    #[test]
-    fn observed_directory_uses_dir_mask() {
-        assert_eq!(
-            resolve_mask_kind("/var/lib/app", None, Some(ObservedMaskPathKind::Dir)).unwrap(),
-            MaskKind::Dir
-        );
-    }
-
-    #[test]
-    fn missing_path_without_explicit_kind_errors() {
-        let err = resolve_mask_kind("/missing/file", None, None).unwrap_err();
-        assert!(err.contains("does not exist"));
-        assert!(err.contains("type"));
-    }
-
-    #[test]
-    fn observed_mask_path_kind_reports_host_dir_file_and_missing() {
-        use std::fs;
-
-        let base = std::env::temp_dir().join(format!(
-            "mxc_observed_mask_kind_{}_{:?}",
-            std::process::id(),
-            std::thread::current().id()
-        ));
-        let dir = base.join("a_dir");
-        let file = base.join("a_file");
-        fs::create_dir_all(&dir).unwrap();
-        fs::write(&file, b"x").unwrap();
-
-        assert_eq!(
-            observed_mask_path_kind(&dir).unwrap(),
-            Some(ObservedMaskPathKind::Dir)
-        );
-        assert_eq!(
-            observed_mask_path_kind(&file).unwrap(),
-            Some(ObservedMaskPathKind::File)
-        );
-        assert_eq!(
-            observed_mask_path_kind(&base.join("does_not_exist")).unwrap(),
-            None
-        );
-
-        fs::remove_dir_all(&base).ok();
-    }
-
-    #[test]
-    fn lookup_mask_kind_prefers_exact_match() {
-        let mut kinds = HashMap::new();
-        kinds.insert("/data/secret".to_string(), MaskKind::File);
-        assert_eq!(
-            lookup_mask_kind(&kinds, "/data/secret"),
-            Some(MaskKind::File)
-        );
-    }
-
-    #[test]
-    fn lookup_mask_kind_tolerates_trailing_separator_mismatch() {
-        // Configured key lacks the trailing slash the resolved mount path carries.
-        let mut kinds = HashMap::new();
-        kinds.insert("/data/secret".to_string(), MaskKind::Dir);
-        assert_eq!(
-            lookup_mask_kind(&kinds, "/data/secret/"),
-            Some(MaskKind::Dir)
-        );
-
-        // Configured key carries a trailing slash the resolved mount path lacks.
-        let mut kinds = HashMap::new();
-        kinds.insert("/data/secret/".to_string(), MaskKind::File);
-        assert_eq!(
-            lookup_mask_kind(&kinds, "/data/secret"),
-            Some(MaskKind::File)
-        );
-    }
-
-    #[test]
-    fn lookup_mask_kind_absent_returns_none() {
-        let kinds = HashMap::new();
-        assert_eq!(lookup_mask_kind(&kinds, "/data/secret"), None);
     }
 }
